drivers: led_strip: add B1414 IC support
The Everlight B1414 IC provides control of a single LED through a single
wire. 24 bits of data must be transmitted (8 bit for each RGB color) to
the B1414 IC to configure this LED. The control signal (waveform) must
meet the following requirements:

Each bit is described with a 1.2 us pulse

0 bit: 300 ns high and 900 ns low.
1 bit: 900 ns high and 300 ns low.

Several B141 IC can be daisy chained. The first device uses the first
24 bits of data to configure its LED and pass the remaining data to
the next device, and so on.

This patch adds a led_strip driver for the B1414 IC. This driver uses
a SPI interface to output a compatible control signal.

Signed-off-by: Simon Guinot <simon.guinot@seagate.com>